+++ packages/vite-plugin-index-html/CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+## 2.0.0
+
+Breakchanges: `preserveEntrySignatures` no longer used as default to meet Vite's output format. You should configure `preserveEntrySignatures` explicitly to preserve entry signatures.
+
+- [chore] set an alias to `input`
+
 ## 1.0.0
 
 - [feat] support minify html using html-minifier-terser.
 
-## 0.1。3
+## 0.1.3
 
 - [feat] return original html when non-matches.